Create Email Analysis.py
This is a bit of a silly vanity script that I used to find out how many jobs I applied for since March of last year. While i filtered my emails for terms like "Application", "Thank you for applying" etc, this was not a perfect process and it picked out too many emails. 

After lengthy deliberation, I settled on a bit of a rudimentary method sampling method where I took out 30 random emails and then tried to select which ones were pertaining to job applications. This gave me a ratio of 24/30, which I assumed was constant for the whole dataset. Perhaps in the future I could make my own machine learning model and train it, but it would be a bit of overkill for what is in essence a silly project